move BirthDeathChain to data, remove reading/writing of matrices
