+++ sktime/data/birth_death_chain.py
