+++ sktime/markov/tools/util/__init__.py
@@ -14,4 +14,3 @@
 #
 # You should have received a copy of the GNU Lesser General Public License
 # along with this program.  If not, see <http://www.gnu.org/licenses/>.
-
