--- sktime/markov/tools/util/annotators.py
@@ -1,48 +1,0 @@
-# This file is part of scikit-time, msmtools, and PyEMMA.
-#
-# Copyright (c) 2020, 2015, 2014 AI4Science Group, Freie Universitaet Berlin (GER)
-#
-# PyEMMA is free software: you can redistribute it and/or modify
-# it under the terms of the GNU Lesser General Public License as published by
-# the Free Software Foundation, either version 3 of the License, or
-# (at your option) any later version.
-#
-# This program is distributed in the hope that it will be useful,
-# but WITHOUT ANY WARRANTY; without even the implied warranty of
-# MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
-# GNU General Public License for more details.
-#
-# You should have received a copy of the GNU Lesser General Public License
-# along with this program.  If not, see <http://www.gnu.org/licenses/>.
-
-__all__ = ['shortcut']
-
-
-def shortcut(*names):
-    """Add an shortcut (alias) to a decorated function, but not to class methods!
-
-    Use aliased/alias decorators for class members!
-
-    Calling the shortcut (alias) will call the decorated function. The shortcut name will be appended
-    to the module's __all__ variable and the shortcut function will inherit the function's docstring
-
-    Examples
-    --------
-    In some module you have defined a function
-    >>> @shortcut('is_tmatrix') # doctest: +SKIP
-    >>> def is_transition_matrix(args): # doctest: +SKIP
-    ...     pass # doctest: +SKIP
-    Now you are able to call the function under its short name
-    >>> is_tmatrix(args) # doctest: +SKIP
-
-    """
-
-    def wrap(f):
-        globals_ = f.__globals__
-        for name in names:
-            globals_[name] = f
-            if '__all__' in globals_ and name not in globals_['__all__']:
-                globals_['__all__'].append(name)
-        return f
-
-    return wrap
