+++ sktime/markov/tools/util/types.py
@@ -428,43 +428,3 @@ def ensure_ndarray_or_None(A, shape=None, uniform=None, ndim=None, size=None, dt
         return ensure_ndarray(A, shape=shape, uniform=uniform, ndim=ndim, size=size, dtype=dtype, kind=kind)
     else:
         return None
-
-
-# ======================================================================================================================
-# EMMA TRAJECTORY TYPES
-# ======================================================================================================================
-
-def ensure_traj(traj):
-    r"""Makes sure that dtraj is a discrete trajectory (array of float)
-
-    """
-    if is_float_matrix(traj):
-        return traj
-    elif is_float_vector(traj):
-        return traj[:,None]
-    else:
-        try:
-            arr = np.array(traj)
-            arr = ensure_dtype_float(arr)
-            if is_float_matrix(arr):
-                return arr
-            if is_float_vector(arr):
-                return arr[:,None]
-            else:
-                raise TypeError('Argument traj cannot be cast into a two-dimensional array. Check type.')
-        except:
-            raise TypeError('Argument traj is not a trajectory - only float-arrays or list of float-arrays are allowed. Check type.')
-
-def ensure_traj_list(trajs):
-    if isinstance(trajs, list):
-        # elements are ints? make it a matrix and wrap into a list
-        if is_list_of_float(trajs):
-            return [np.array(trajs)[:,None]]
-        else:
-            res = []
-            for i in range(len(trajs)):
-                res.append(ensure_traj(trajs[i]))
-            return res
-    else:
-        # looks like this is one trajectory
-        return [ensure_traj(trajs)]
