+++ tests/markov/tools/analysis/impl/dense/assessment_test.py
@@ -22,8 +22,8 @@ Created on 07.10.2013
 """
 import unittest
 import numpy as np
-from sktime.markov.tools.util.birth_death_chain import BirthDeathChain
 
+from sktime.data.birth_death_chain import BirthDeathChain
 from sktime.markov.tools.analysis.dense import assessment
