+++ tests/markov/tools/analysis/impl/dense/committor_test.py
@@ -23,7 +23,8 @@ r"""Unit tests for the committor module
 
 import unittest
 import numpy as np
-from sktime.markov.tools.util.birth_death_chain import BirthDeathChain
+
+from sktime.data.birth_death_chain import BirthDeathChain
 from tests.markov.tools.numeric import assert_allclose
 
 from sktime.markov.tools.analysis.dense import committor
