+++ tests/markov/tools/analysis/impl/dense/correlations_test.py
@@ -22,7 +22,8 @@ Created on Jun 3, 2014
 """
 import unittest
 import numpy as np
-from sktime.markov.tools.util.birth_death_chain import BirthDeathChain
+
+from sktime.data.birth_death_chain import BirthDeathChain
 from tests.markov.tools.numeric import assert_allclose
 
 from sktime.markov.tools.analysis.dense import correlations
