+++ tests/markov/tools/analysis/impl/dense/decomposition_test.py
@@ -25,7 +25,8 @@ import unittest
 import warnings
 
 import numpy as np
-from sktime.markov.tools.util.birth_death_chain import BirthDeathChain
+
+from sktime.data.birth_death_chain import BirthDeathChain
 from tests.markov.tools.numeric import assert_allclose
 
 from scipy.linalg import eigvals
