+++ tests/markov/tools/analysis/impl/dense/stationary_vector_test.py
@@ -18,7 +18,8 @@
 
 import unittest
 import numpy as np
-from sktime.markov.tools.util.birth_death_chain import BirthDeathChain
+
+from sktime.data.birth_death_chain import BirthDeathChain
 from tests.markov.tools.numeric import assert_allclose
 
 from sktime.markov.tools.analysis.dense.stationary_vector import stationary_distribution_from_eigenvector
