+++ tests/markov/tools/analysis/impl/sparse/fingerprints_test.py
@@ -24,7 +24,8 @@ r"""Unit test for the fingerprint module
 import unittest
 
 import numpy as np
-from sktime.markov.tools.util.birth_death_chain import BirthDeathChain
+
+from sktime.data.birth_death_chain import BirthDeathChain
 from tests.markov.tools.numeric import assert_allclose
 
 from sktime.markov.tools.analysis.sparse.decomposition import rdl_decomposition, timescales
