--- tests/markov/tools/util/testfiles/matrix_complex.dat
@@ -1,3 +1,0 @@
- (0.000000000000000000e+00+9.000000000000000000e+00j)  (1.000000000000000000e+00+1.000000000000000000e+01j)  (2.000000000000000000e+00+1.100000000000000000e+01j)
- (3.000000000000000000e+00+1.200000000000000000e+01j)  (4.000000000000000000e+00+1.300000000000000000e+01j)  (5.000000000000000000e+00+1.400000000000000000e+01j)
- (6.000000000000000000e+00+1.500000000000000000e+01j)  (7.000000000000000000e+00+1.600000000000000000e+01j)  (8.000000000000000000e+00+1.700000000000000000e+01j)
