--- tests/markov/tools/util/testfiles/spmatrix_float.coo.dat
@@ -1,27 +1,0 @@
-1.000000000000000000e+00 1.000000000000000000e+00 1.000000000000000000e+00
-2.000000000000000000e+00 2.000000000000000000e+00 2.000000000000000000e+00
-3.000000000000000000e+00 3.000000000000000000e+00 3.000000000000000000e+00
-4.000000000000000000e+00 4.000000000000000000e+00 4.000000000000000000e+00
-5.000000000000000000e+00 5.000000000000000000e+00 5.000000000000000000e+00
-6.000000000000000000e+00 6.000000000000000000e+00 6.000000000000000000e+00
-7.000000000000000000e+00 7.000000000000000000e+00 7.000000000000000000e+00
-8.000000000000000000e+00 8.000000000000000000e+00 8.000000000000000000e+00
-9.000000000000000000e+00 9.000000000000000000e+00 9.000000000000000000e+00
-0.000000000000000000e+00 1.000000000000000000e+00 1.000000000000000000e+01
-1.000000000000000000e+00 2.000000000000000000e+00 1.100000000000000000e+01
-2.000000000000000000e+00 3.000000000000000000e+00 1.200000000000000000e+01
-3.000000000000000000e+00 4.000000000000000000e+00 1.300000000000000000e+01
-4.000000000000000000e+00 5.000000000000000000e+00 1.400000000000000000e+01
-5.000000000000000000e+00 6.000000000000000000e+00 1.500000000000000000e+01
-6.000000000000000000e+00 7.000000000000000000e+00 1.600000000000000000e+01
-7.000000000000000000e+00 8.000000000000000000e+00 1.700000000000000000e+01
-8.000000000000000000e+00 9.000000000000000000e+00 1.800000000000000000e+01
-1.000000000000000000e+00 0.000000000000000000e+00 2.000000000000000000e+01
-2.000000000000000000e+00 1.000000000000000000e+00 2.100000000000000000e+01
-3.000000000000000000e+00 2.000000000000000000e+00 2.200000000000000000e+01
-4.000000000000000000e+00 3.000000000000000000e+00 2.300000000000000000e+01
-5.000000000000000000e+00 4.000000000000000000e+00 2.400000000000000000e+01
-6.000000000000000000e+00 5.000000000000000000e+00 2.500000000000000000e+01
-7.000000000000000000e+00 6.000000000000000000e+00 2.600000000000000000e+01
-8.000000000000000000e+00 7.000000000000000000e+00 2.700000000000000000e+01
-9.000000000000000000e+00 8.000000000000000000e+00 2.800000000000000000e+01
