--- tests/markov/tools/util/testfiles/spmatrix_int_reference.dat
@@ -1,10 +1,0 @@
-0 10 0 0 0 0 0 0 0 0
-20 1 11 0 0 0 0 0 0 0
-0 21 2 12 0 0 0 0 0 0
-0 0 22 3 13 0 0 0 0 0
-0 0 0 23 4 14 0 0 0 0
-0 0 0 0 24 5 15 0 0 0
-0 0 0 0 0 25 6 16 0 0
-0 0 0 0 0 0 26 7 17 0
-0 0 0 0 0 0 0 27 8 18
-0 0 0 0 0 0 0 0 28 9
